Adds endpoint for Admin login
